+++ pojo-lens/src/main/java/laughing/man/commits/telemetry/QueryTelemetryStage.java
@@ -6,6 +6,7 @@ package laughing.man.commits.telemetry;
 public enum QueryTelemetryStage {
     PARSE,
     BIND,
+    PUSHDOWN,
     FILTER,
     AGGREGATE,
     ORDER,